pyproject.toml
# =============================================================================
# uSipipo VPN Bot - Project Configuration
# =============================================================================
# Python: >=3.13
# Build System: uv/hatchling
# =============================================================================

[project]
name = "usipipobot"
version = "3.4.0"
description = "VPN Manager Telegram Bot with Mini App"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "uSipipo Team", email = "dev@usipipo.com"}
]
keywords = ["vpn", "telegram", "bot", "wireguard", "outline"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
]

# Runtime dependencies
dependencies = [
    # Telegram Bot
    "python-telegram-bot[job-queue]==21.10",

    # Database (PostgreSQL + SQLAlchemy Async)
    "sqlalchemy[asyncio]==2.0.48",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    "psycopg2-binary==2.9.10",

    # Configuration & Validation
    "pydantic==2.12.5",
    "pydantic-settings==2.6.1",
    "python-dotenv==1.2.2",
    "PyYAML==6.0.3",

    # HTTP Client
    "httpx==0.28.1",
    "aiohttp==3.13.3",
    "redis==7.3.0",

    # FastAPI for Webhook API
    "fastapi==0.135.1",
    "uvicorn[standard]==0.41.0",
    "python-multipart==0.0.22",
    "jinja2==3.1.6",
    "itsdangerous==2.2.0",

    # Security & Authentication
    "PyJWT==2.12.0",
    "cryptography==46.0.5",

    # Utilities
    "punq==0.7.0",
    "loguru==0.7.3",
    "qrcode[pil]==8.2",
    "pillow==12.1.1",
    "cachetools==7.0.5",
    "python-dateutil==2.8.2",
    "pytz==2026.1.post1",
    "psutil==7.1.0",

    # Optional / Development
    "rich==14.2.0",
    "StrEnum==0.4.15",
    "tenacity==9.1.4",
]

[project.optional-dependencies]
dev = [
    "pre-commit==3.6.0",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov>=4.0.0",
    "flake8",
    "black",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "application",
    "config",
    "domain",
    "infrastructure",
    "miniapp",
    "telegram_bot",
    "utils",
]

# =============================================================================
# UV Configuration
# =============================================================================

[dependency-groups]
dev = [
    "pre-commit==3.6.0",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov>=4.0.0",
    "flake8",
    "black",
    "mypy",
    "isort>=8.0.1",
    "detect-secrets>=1.5.0",
]

# =============================================================================
# Pytest Configuration
# =============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "asyncio: mark test as async",
    "integration: mark test as integration test",
    "slow: mark test as slow test",
]

# =============================================================================
# Black Configuration
# =============================================================================

[tool.black]
line-length = 100
target-version = ["py313"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | \.eggs
  | \.mypy_cache
  | \.pytest_cache
  | build
  | dist
)/
'''

# =============================================================================
# isort Configuration
# =============================================================================

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
known_first_party = ["domain", "application", "infrastructure", "telegram_bot", "miniapp", "utils", "config"]
known_third_party = ["pytest", "pydantic", "sqlalchemy", "fastapi", "telegram", "httpx", "aiohttp"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = [".venv", "venv", ".git", "migrations/versions"]

# =============================================================================
# Mypy Configuration
# =============================================================================

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false
